pyproject.toml
[project]
name = "learn-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.0.1",
    "langchain-openai>=1.0.1",
    "langchain-openapi>=0.1.1",
    "langchain[anthropic]>=1.0.2",
    "langgraph>=1.0.1",
    "python-dotenv>=1.1.1",
]
